Extract data from RUIAN

Extract specific layer, in form of spatial data, from RUIAN for given settlement in Czech Republic. Checks are performed to find out if the provided id is valid for some settlement in Czech Republic.

load_RUIAN_settlement(id, layer = "obec", WGS84 = FALSE)

Arguments

id

id of settlement as character

layer

identification of data to extract as character, see details. Default value is "obec"

WGS84

convert data to WGS-84 coordinate system? Default FALSE.

Value

data.frame with spatial objects (sf) of the specified layer

Details

In case of adress places (using ADRM_B or adresni mista as layer) are checked, the csv file with more attributes is also downloaded and linked to the spatial data layer.

The layer can have values from following set, the value in brackets is alias to full layer name:

"ADRM_B" ("adresni mista")

"CO_B" ("casti obce")

"KATUZE_P" ("katastralni uzemi")

"OBEC_P" ("obec")

"SO_B" ("stavebni objekty")

"UL_L" ("ulice")

"VO_P" ("volebni okrsky")

"ZSJ_P" ("zakladni sidelni jednotky")

"MOMC_P"

"MOP_P"

"SOP_P"

So the codes layer = "CO_B" and layer = "casti obce" are equal.

The values of id follow general pattern of six number with first number being 5.

Information about dataset

More detailed information about data can be found at the provider's website http://atom.cuzk.cz/.

Examples

if (FALSE) {
    adresy_vyskov <- load_RUIAN_settlement("592889", layer = "adresni mista")
}