Extract data from Cadastral map

Extract specific layer, in form of spatial data, from cadastral map for given cadastral territory in Czech Republic. Checks are performed to find out if the provided id is valid for some cadastral territory in Czech Republic.

load_cadastral_territory(id, layer = "katastralni uzemi", WGS84 = FALSE)

Arguments

id

id of cadastral territory as character

layer

identification of data to extract as character, see details. Default value is "katastralni území"

WGS84

convert data to WGS-84 coordinate system? Default FALSE.

Value

data.frame with spatail objects (sf) of the specified layer

Details

The layer can have values from following set, the value in brackets is alias to full layer name:

"BODOVE_POLE_B"

"BODOVE_POLE_T"

"BUDOVY_B" ("budovy body")

"BUDOVY_DEF"

"BUDOVY_P" ("budovy")

"DALSI_PRVKY_MAPY_B"

"DALSI_PRVKY_MAPY_L"

"DALSI_PRVKY_MAPY_T"

"HRANICE_PARCEL_L" ("hranice parcel")

"KATASTRALNI_UZEMI_DEF"

"KATASTRALNI_UZEMI_L"

"KATASTRALNI_UZEMI_P" ("katastralni uzemi")

"PARCELY_KN_B"

"PARCELY_KN_DEF"

"PARCELY_KN_L"

"PARCELY_KN_P" ("parcely")

"PRVKY_ORIENT_MAPY_B"

"PRVKY_ORIENT_MAPY_L"

"PRVKY_ORIENT_MAPY_T"

"VB_P"

So the codes layer = "BUDOVY_B" and layer = "budovy body" are equal.

The values of id follow general pattern of six number with first number being 6,7 or 9.

Information about dataset

More detailed information about data can be found at the provider's website http://atom.cuzk.cz/.

Examples

if (FALSE) {
    parcely_vyskov <- load_cadastral_territory("788571", layer = "parcely")
}