Load or save from Data200

Load data from Data200 data source (https://geoportal.cuzk.cz/(S(ijginumejzilvacbfijkylwj))/Default.aspx?mode=TextMeta&side=mapy_data200&text=dSady_mapyData200&head_tab=sekce-02-gp&menu=229). The data can be used only after correctly citing the creator (as per terms of use https://geoportal.cuzk.cz/Dokumenty/Podminky_EN.pdf). The citation is in form "Mapový podklad – Data200, insert year © Český úřad zeměměřický a katastrální, www.cuzk.cz".

load_Data200(layer, WGS84 = FALSE)

save_Data200(path, layer = NULL, type = NULL)

load_Data200_info(english_names = FALSE)

Arguments

layer

identification of data to extract as character, see details.

WGS84

convert data to WGS-84 coordinate system? Default FALSE.

path

character path to store the files to.

type

character type of layers to save. See details, types are listed in brackets.

english_names

change the names of the columns to English. Default FALSE.

Value

"load_Data200" - data.frame with spatial objects (sf) of the specified layer. For layer either "DMR" or "DMRShaded" the output is actually a (raster). "save_Data200" - path to the unzipped files (for layer) or folder (for type), the zipped file is also stored at path (mainly for further use)

Details

The layer can have values from following set, in the bracket is the name of general category (can be used as type in saving the data):

"AdministrativniHraniceLinie" ("Hranice")

"AdministrativniUzemiCentroid" ("Hranice")

"AdministrativniUzemiUTJ" ("Hranice")

"AdministrativniUzemiObce" ("Hranice")

"AdministrativniUzemiOkresy" ("Hranice")

"AdministrativniUzemiKraje" ("Hranice")

"HrazJezNad50m" ("Vodstvo")

"HrazJezPod50m" ("Vodstvo")

"JezeroRybnikVodniNadrz" ("Vodstvo")

"VodniTokPod50m" ("Vodstvo")

"VodniTokNad50m" ("Vodstvo")

"Ostrovy" ("Vodstvo")

"MokrinaBazina" ("Vodstvo")

"Vodopad" ("Vodstvo")

"Prameny1" ("Vodstvo")

"Prameny2" ("Vodstvo")

"OrografickeNazvy" ("Popis")

"GeomorfologickeOblasti" ("Popis")

"GeomorfologickeCelky" ("Popis")

"GeomorfologickePodcelky" ("Popis")

"NarodniParkPrirodniRezervace" ("RuzneObjekty")

"Produktovod" ("RuzneObjekty")

"Vysilac" ("RuzneObjekty")

"VyznamneObjekty" ("RuzneObjekty")

"ProduktovodVyznamneBody" ("RuzneObjekty")

"Věž" ("RuzneObjekty")

"DulLom" ("RuzneObjekty")

"Budova" ("RuzneObjekty")

"ElektrickeVedeni" ("RuzneObjekty")

"Elektrarna" ("RuzneObjekty")

"ObceBody" ("Sidla")

"ObcePolygony" ("Sidla")

"Privoz" ("Doprava")

"PrivozStanice" ("Doprava")

"ZeleznicniPrejezd" ("Doprava")

"Heliport" ("Doprava")

"LanovaDraha" ("Doprava")

"DalnicniOdpocivka" ("Doprava")

"KrizovatkaMimourovnova" ("Doprava")

"LetisteNad40Ha" ("Doprava")

"LetisteNad40HaBod" ("Doprava")

"ZelezniceZastavky" ("Doprava")

"LetistePod40Ha" ("Doprava")

"LodniPristav" ("Doprava")

"PristavaciDraha" ("Doprava")

"Zeleznice" ("Doprava")

"Silnice" ("Doprava")

"LesyPlantaze" ("Vegetace")

"KotovaneBody" ("Relief")

"Vrstevnice" ("Relief")

"SkalniStenaSraz" ("Relief")

"Jeskyne" ("Relief")

"DMR" ("Relief")

"DMRShaded" ("Relief")

Functions

load_Data200: Loads single dataset

save_Data200: Download and store layer (and zipped general category) or complete category

load_Data200_info: Load information about layers in Data200.

Examples

if (FALSE) {
waterfalls <- load_Data200(layer = "Vodopad")
}
if (FALSE) {
folder_water_objects <- save_Data200("~/data/water", type = "Vodopad")
}