Load or save from Data50

Load data from Data50 data source (https://geoportal.cuzk.cz/(S(xbw0cmgh1cve4bciko2oo4e2))/Default.aspx?lng=EN&mode=TextMeta&side=mapy_data50&text=dSady_mapyData50&head_tab=sekce-02-gp&menu=2290). The data can be used only after correctly citing the creator (as per terms of use https://geoportal.cuzk.cz/Dokumenty/Podminky_EN.pdf). The citation is in form "Mapový podklad – Data50, insert year © Český úřad zeměměřický a katastrální, www.cuzk.cz".

Some basic description of the dataset Data50 or Data200. Most importantly names of layers and sizes of files that need to be downloaded.

load_Data50(layer, WGS84 = FALSE)

save_Data50(path, layer = NULL, type = NULL)

load_Data50_info(english_names = FALSE)

Arguments

layer

identification of data to extract as character, see details.

WGS84

convert data to WGS-84 coordinate system? Default FALSE.

path

character path to store the files to.

type

character type of layers to save. See details, types are listed in brackets.

english_names

change the names of the columns to English. Default FALSE.

Value

"load_Data50" - data.frame with spatial objects (sf) of the specified layer. "save_Data50" - path to the unzipped files (for layer) or folder (for type), the zipped file is also stored at path (mainly for further use) data.frame with description of layers.

Details

The layer can have values from following set, in the bracket is the name of general category (can be used as type in saving the data):

"BlokBudov" ("sidelniKulturniHospodarskeObjekty")

"Budova" ("sidelniKulturniHospodarskeObjekty")

"Hrad" ("sidelniKulturniHospodarskeObjekty")

"Hrbitov" ("sidelniKulturniHospodarskeObjekty")

"ChatovaKolonie" ("sidelniKulturniHospodarskeObjekty")

"Kostel" ("sidelniKulturniHospodarskeObjekty")

"LyzarskyMustek" ("sidelniKulturniHospodarskeObjekty")

"RozhlednaVysilac" ("sidelniKulturniHospodarskeObjekty")

"Rozvalina" ("sidelniKulturniHospodarskeObjekty")

"Stadion" ("sidelniKulturniHospodarskeObjekty")

"UsazovaciNadrzOdkaliste" ("sidelniKulturniHospodarskeObjekty")

"VetrnyMotor" ("sidelniKulturniHospodarskeObjekty")

"VezovitaStavba" ("sidelniKulturniHospodarskeObjekty")

"VodojemVezovy" ("sidelniKulturniHospodarskeObjekty")

"Zamek" ("sidelniKulturniHospodarskeObjekty")

"Zricenina" ("sidelniKulturniHospodarskeObjekty")

"Cesta" ("komunikace")

"LanovaDraha" ("komunikace")

"Lavka" ("komunikace")

"Letiste" ("komunikace")

"LetisteObvodovaLinie" ("komunikace")

"Most" ("komunikace")

"Pesina" ("komunikace")

"Pristav" ("komunikace")

"Privoz" ("komunikace")

"SilniceDalnice" ("komunikace")

"SilniceVeVystavbe" ("komunikace")

"Tunel" ("komunikace")

"Ulice" ("komunikace")

"ZeleznicniStanice" ("komunikace")

"ZeleznicniTrat" ("komunikace")

"ZeleznicniVlecka" ("komunikace")

"ElektrickeVedeni" ("produktovodyElektrickeVedeni")

"Produktovod" ("produktovodyElektrickeVedeni")

"Akvadukt" ("vodstvo")

"Hraz" ("vodstvo")

"Jez" ("vodstvo")

"Shybka" ("vodstvo")

"VodniPlocha" ("vodstvo")

"VodniTok" ("vodstvo")

"HraniceSpravniJednotkyaKU" ("hraniceUzemnichJednotek")

"ChraneneUzemi" ("hraniceUzemnichJednotek")

"Les" ("vegetacePovrch")

"LoukaPastvina" ("vegetacePovrch")

"RaselinisteMocalBazina" ("vegetacePovrch")

"ZahradaSadParkViniceChmelnice" ("vegetacePovrch")

"Jeskyne" ("terenniRelief")

"KotovanyBod" ("terenniRelief")

"SkalnatySraz" ("terenniRelief")

"Skaly" ("terenniRelief")

"TerenniStupen" ("terenniRelief")

"TerenniStupenSpadnice" ("terenniRelief")

"Vrstevnice" ("terenniRelief")

"DefinicniBodCastiObce" ("popis")

"DefinicniBodSpravnihoCelku" ("popis")

"Jmeno_B" ("popis")

"Jmeno_L" ("popis")

"Jmeno_P" ("popis")

"ObjektRuzny" ("popis")

Functions

load_Data50: Loads single dataset

save_Data50: Download and store layer (and zipped general category) or complete category

load_Data50_info: Load information about layers in Data50.

Examples

if (FALSE) {
    rivers <- load_Data50(layer = "VodniTok")
}
if (FALSE) {
    folder_communications <- save_Data50("~/data/coomunications", type = "komunikace")
}
if (FALSE) {
    info <- load_Data50_info(english_names = TRUE)
    info <- load_Data200_info(english_names = TRUE)
}